+++ landingzones/forms.py
@@ -107,7 +107,10 @@ class LandingZoneForm(forms.ModelForm):
                         )
             # Disable assay dropdown if only one assay is availalbe
             if len(self.fields['assay'].widget.choices) == 1:
-                self.fields['assay'].widget.attrs['disabled'] = True
+                self.fields['assay'].disabled = True
+                self.initial['assay'] = self.fields['assay'].widget.choices[0][
+                    0
+                ]
             # Set options and initial value for coll_creation
             self.fields['coll_creation'].widget = forms.Select()
             self.fields[